Remove `printError()`, Finish AppTest.java Tests, All Tests Pass
All 40 tests pass.

`App.printError(String)` is removed as it is hard to test using JUnit and it doesn't foster abstraction (the lines can be moved outside the function without notably increasing the program size, and doing so reduces an unnecessary function call).

`App.printStationUptimes(int[][])` will not be tested using JUnit, since it prints onto the console in a straightforward manner.

The above two changes render the new PrintStreams unnecessary.

Replaced `getMainFilePath(String)` with `getRelativeFilePath(String)`, the behavior of which depends on the value of `RUNNING_FROM_SRC_SUBDIRECTORY`. See documentation and the README for details.